requirements.txt
# requirements.txt

aiohttp~=3.10.5
asyncpg~=0.29.0
coloredlogs==15.0.1
Hypercorn==0.17.3
python-dotenv==1.0.1
python-socketio~=5.11.4
Quart~=0.19.6
Werkzeug~=3.0.4
discord-webhook[async]==1.3.1
geoip2==4.8.1

pytest~=8.3.4
requests~=2.32.3
javaobj-py3~=0.4.4